+++ AddOns/RaiderIO/db/db_recruitment_cn_characters.lua
@@ -2,11 +2,13 @@
 -- Copyright (c) 2026 by RaiderIO, Inc.
 -- All rights reserved.
 --
-local provider={name=...,data=3,region="cn",date="2026-04-18T07:34:57Z",numCharacters=34413,db={}}
+local provider={name=...,data=3,region="cn",date="2026-04-19T07:33:50Z",numCharacters=34478,db={}}
 local F
 
 F = function() provider.db["无尽之海"]={0,"Greyghost","Moondance","Pen","Tenpen","一点二十丸","丨释怀","乄德","乔乔小弟","九点二十一","可怜德古丹","周哲伦","大废止者","整活大師","札皮","渣皮","無料案内人","独轶望江楼","电击治疗丶","章鱼哥的邻居","轧皮","释丨怀","阿蒙内特","骑骑丶"} end F()
-F = function() provider.db["米奈希尔"]={46,"章鱼术"} end F()
+F = function() provider.db["伊森利恩"]={46,"劍灬來"} end F()
+F = function() provider.db["白银之手"]={48,"阿渝子"} end F()
+F = function() provider.db["米奈希尔"]={50,"章鱼术"} end F()
 
 F = nil
 RaiderIO.AddProvider(provider)