+++ AddOns/RaiderIO/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Raider.IO Mythic Plus, Raiding, and Recruitment
 
-## [v202604180600](https://github.com/RaiderIO/raiderio-addon/tree/v202604180600) (2026-04-18)
-[Full Changelog](https://github.com/RaiderIO/raiderio-addon/compare/v202604170600...v202604180600) [Previous Releases](https://github.com/RaiderIO/raiderio-addon/releases)
+## [v202604190600](https://github.com/RaiderIO/raiderio-addon/tree/v202604190600) (2026-04-19)
+[Full Changelog](https://github.com/RaiderIO/raiderio-addon/compare/v202604180600...v202604190600) [Previous Releases](https://github.com/RaiderIO/raiderio-addon/releases)
 
 - [Raider.IO] Database Refresh  
 - [Raider.IO] Classic Database Refresh